+++ packages/jsii-pacmak/test/expected.jsii-calc-lib/python/src/scope/jsii_calc_lib/__init__.py
@@ -10,7 +10,7 @@ import publication
 
 import scope.jsii_calc_base
 
-__jsii_assembly__ = jsii.JSIIAssembly.load("@scope/jsii-calc-lib", "0.20.11", __name__, "jsii-calc-lib@0.20.11.jsii.tgz")
+__jsii_assembly__ = jsii.JSIIAssembly.load("@scope/jsii-calc-lib", "0.21.1", __name__, "jsii-calc-lib@0.21.1.jsii.tgz")
 
 
 @jsii.enum(jsii_type="@scope/jsii-calc-lib.EnumFromScopedModule")