+++ packages/@jsii/ruby-runtime/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.21.1](https://github.com/aws/jsii/compare/v0.21.0...v0.21.1) (2020-01-03)
+
+**Note:** Version bump only for package @jsii/ruby-runtime
+
+
+
+
+
+# [0.21.0](https://github.com/aws/jsii/compare/v0.20.11...v0.21.0) (2020-01-02)
+
+**Note:** Version bump only for package @jsii/ruby-runtime
+
+
+
+
+
 ## [0.20.11](https://github.com/aws/jsii/compare/v0.20.10...v0.20.11) (2019-12-13)
 
 **Note:** Version bump only for package @jsii/ruby-runtime